# backend/openapi-spec.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "SkillBridge API Ecosystem",
    "description": "Skill Intelligence and Career Readiness Platform API Specifications",
    "contact": {
      "name": "SkillBridge Development Team",
      "email": "support@skillbridge.internal"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://springdoc.org"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/careers/target": {
      "post": {
        "tags": [
          "Career Intelligence Service"
        ],
        "summary": "Set target career role for student profile",
        "operationId": "selectTargetRole",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TargetRoleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseStudentProfile"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/register": {
      "post": {
        "tags": [
          "Authentication Service"
        ],
        "summary": "Register a new student user",
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseAuthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/refresh": {
      "post": {
        "tags": [
          "Authentication Service"
        ],
        "summary": "Refresh JWT authentication token",
        "operationId": "refresh",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseAuthResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/login": {
      "post": {
        "tags": [
          "Authentication Service"
        ],
        "summary": "Authenticate user and issue JWT token",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseAuthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/forgot-password": {
      "post": {
        "tags": [
          "Authentication Service"
        ],
        "summary": "Initiate password reset request (Mocked for development)",
        "operationId": "forgotPassword",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseString"
                }
              }
            }
          }
        }
      }
    },
    "/api/assessments/{id}/submit": {
      "post": {
        "tags": [
          "Assessment Service"
        ],
        "summary": "Submit assessment responses for score evaluation and skill verification",
        "operationId": "submitAssessment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssessmentSubmissionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseAssessmentResultDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/assessments/{id}/start": {
      "post": {
        "tags": [
          "Assessment Service"
        ],
        "summary": "Start assessment test session",
        "operationId": "startAssessment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseMapStringObject"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/careers/{roleId}": {
      "get": {
        "tags": [
          "Career Intelligence Service"
        ],
        "summary": "Get specific career role details",
        "operationId": "getCareer",
        "parameters": [
          {
            "name": "roleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseCareerRole"
                }
              }
            }
          }
        }
      }
    },
    "/api/careers/{roleId}/roadmap": {
      "get": {
        "tags": [
          "Career Intelligence Service"
        ],
        "summary": "Get structured career development roadmap",
        "operationId": "getRoadmap",
        "parameters": [
          {
            "name": "roleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseCareerRoadmapDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/careers/{roleId}/competencies": {
      "get": {
        "tags": [
          "Career Intelligence Service"
        ],
        "summary": "Get required competencies for a target role",
        "operationId": "getCompetencies",
        "parameters": [
          {
            "name": "roleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListCareerCompetency"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/assessments/{id}": {
      "get": {
        "tags": [
          "Assessment Service"
        ],
        "summary": "Get specific assessment details",
        "operationId": "getAssessmentById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseAssessment"
                }
              }
            }
          }
        }
      }
    },
    "/api/assessments/{id}/questions": {
      "get": {
        "tags": [
          "Assessment Service"
        ],
        "summary": "Get questions for an assessment",
        "operationId": "getQuestions",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListAssessmentQuestion"
                }
              }
            }
          }
        }
      }
    },
    "/api/assessments/results/{resultId}": {
      "get": {
        "tags": [
          "Assessment Service"
        ],
        "summary": "Get detailed result for an assessment attempt",
        "operationId": "getResult",
        "parameters": [
          {
            "name": "resultId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseAssessmentAttempt"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AssessmentAttempt": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "assessmentId": {
            "type": "integer",
            "format": "int64"
          },
          "assessmentTitle": {
            "type": "string"
          },
          "skillId": {
            "type": "string"
          },
          "score": {
            "type": "integer",
            "format": "int32"
          },
          "passed": {
            "type": "boolean"
          },
          "competencyLevel": {
            "type": "string"
          },
          "submittedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "StudentProfile": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "fullName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "education": {
            "type": "string"
          },
          "experienceLevel": {
            "type": "string"
          },
          "targetRoleId": {
            "type": "string"
          },
          "targetRoleTitle": {
            "type": "string"
          },
          "resumeUrl": {
            "type": "string"
          },
          "certificatesJson": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "ApiResponseString": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "errorCode": {
            "type": "string"
          }
        }
      },
...
      "TargetRoleRequest": {
        "required": [
          "roleId"
        ],
        "type": "object",
        "properties": {
          "roleId": {
            "type": "string"
          }
        }
      },
      "ApiResponseStudentProfile": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/StudentProfile"
          },
          "message": {
            "type": "string"
          },
          "errorCode": {
            "type": "string"
          }
        }
      },
      "RegisterRequest": {
        "required": [
          "email",
          "fullName",
          "password"
        ],
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "maxLength": 2147483647,
            "minLength": 6,
            "type": "string"
          }
        }
      },
      "ApiResponseAuthResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/AuthResponse"
          },
          "message": {
            "type": "string"
          },
          "errorCode": {
            "type": "string"
          }
        }
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "fullName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "token": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/UserDto"
          }
        }
      },
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "fullName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "avatar": {
            "type": "string"
          },
          "targetRole": {
            "type": "string"
          }
        }
      },
      "LoginRequest": {
        "required": [
          "email",
          "password"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "ForgotPasswordRequest": {
        "required": [
          "email"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          }
        }
      },
      "AssessmentSubmissionRequest": {
        "type": "object",
        "properties": {
          "answers": {
            "type": "object",
            "additionalProperties": {
              "type": "integer",
              "format": "int32"
            }
          }
        }
      },
      "ApiResponseAssessmentResultDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/AssessmentResultDto"
          },
          "message": {
            "type": "string"
          },
          "errorCode": {
            "type": "string"
          }
        }
      },
      "AssessmentResultDto": {
        "type": "object",
        "properties": {
          "attempt": {
            "$ref": "#/components/schemas/AssessmentAttempt"
          },
          "skillVerificationStatus": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "ApiResponseMapStringObject": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "object",
            "additionalProperties": {
              "type": "object"
            }
          },
          "message": {
            "type": "string"
          },
          "errorCode": {
            "type": "string"
          }
        }
      },
...
      "CareerRole": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "roleId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "importance": {
            "type": "string"
          },
          "requiredSkillsCount": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "ApiResponseCareerRole": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/CareerRole"
          },
          "message": {
            "type": "string"
          },
          "errorCode": {
            "type": "string"
          }
        }
      },
      "ApiResponseCareerRoadmapDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/CareerRoadmapDto"
          },
          "message": {
            "type": "string"
          },
          "errorCode": {
            "type": "string"
          }
        }
      },
      "CareerCompetency": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "roleId": {
            "type": "string"
          },
          "skillId": {
            "type": "string"
          },
          "skillName": {
            "type": "string"
          },
          "requiredLevel": {
            "type": "string"
          },
          "importanceWeight": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "CareerRoadmapDto": {
        "type": "object",
        "properties": {
          "role": {
            "$ref": "#/components/schemas/CareerRole"
          },
          "competencies": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CareerCompetency"
            }
          },
          "stages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoadmapStage"
            }
          }
        }
      },
      "RoadmapStage": {
        "type": "object",
        "properties": {
          "stageName": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "skills": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "estimatedDuration": {
            "type": "string"
          }
        }
      },
      "ApiResponseListCareerCompetency": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CareerCompetency"
            }
          },
          "message": {
            "type": "string"
          },
          "errorCode": {
            "type": "string"
          }
        }
      },
...
      "Assessment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "skillId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "questionCount": {
            "type": "integer",
            "format": "int32"
          },
          "durationMinutes": {
            "type": "integer",
            "format": "int32"
          },
          "difficulty": {
            "type": "string"
          },
          "passScore": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "ApiResponseAssessment": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/Assessment"
          },
          "message": {
            "type": "string"
          },
          "errorCode": {
            "type": "string"
          }
        }
      },
      "ApiResponseListAssessmentQuestion": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AssessmentQuestion"
            }
          },
          "message": {
            "type": "string"
          },
          "errorCode": {
            "type": "string"
          }
        }
      },
      "AssessmentQuestion": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "assessmentId": {
            "type": "integer",
            "format": "int64"
          },
          "questionText": {
            "type": "string"
          },
          "optionsJson": {
            "type": "string"
          },
          "correctOptionIndex": {
            "type": "integer",
            "format": "int32"
          },
          "explanation": {
            "type": "string"
          }
        }
      },
      "ApiResponseAssessmentAttempt": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/AssessmentAttempt"
          },
          "message": {
            "type": "string"
          },
          "errorCode": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "name": "bearerAuth",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}